Recover trickyFn_8013b368 ((u32)tp != 0 cast -> cmplwi unsigned null-check)

+++ src/main/dll/trickyfollow.c
@@ -266,7 +266,7 @@ int trickyFn_8013b368(u8* obj, f32 vel, u8* state)
         }
         else if (((TrickyState*)state)->unk09 < 5)
         {
-            if (tp != 0)
+            if ((u32)tp != 0)
             {
                 if (targetWg == 0)
                 {
@@ -343,7 +343,7 @@ int trickyFn_8013b368(u8* obj, f32 vel, u8* state)
                         if (wg == 0)
                         {
                             tp = getPatchGroup((f32*)(obj + 0x18), ((TrickyState*)state)->unkD0) & 0xffff;
-                            if (tp != 0)
+                            if ((u32)tp != 0)
                             {
                                 if (*(s16*)&((TrickyState*)state)->unkD2 == tp)
                                 {
